refactor(kentico-digital-experience): polish custom action skill and rename plugin
Rename the automation-implementation plugin to kentico-digital-experience as
part of the KentiCopilot repo restructure, and align the automation-action-create
skill with the current custom automation action API.

- Remove the hardcoded references/automation-customization.md snapshot and point
  references/docs.md and SKILL.md at the live "Custom automation steps"
  documentation fetched via the Kentico Docs MCP.
- Update examples and guidance to the current API: retrieve the contact via
  context.GetProcessedObject, use the RegisterAutomationAction Description
  parameter instead of the unsupported Tooltip, reference Icons.* constants
  instead of raw "xp-" strings, and store the identifier in an IDENTIFIER constant.
- Split each properties class into its own file and document the
  one-class-per-file guardrail.
