+++ README.md
@@ -19,9 +19,9 @@ This repository provides plugins, each containing a set of skills for AI coding 
 
 ### Automation
 
-> **Location:** [plugins/automation-implementation/](./plugins/automation-implementation/)
+> **Location:** [plugins/kentico-digital-experience/](./plugins/kentico-digital-experience/)
 
-AI-assisted implementation of [Automation process](https://docs.kentico.com/documentation/business-users/digital-marketing/automation) extensions in Xperience by Kentico — starting with **custom automation actions** (custom step types in the Automation Builder). The AI reviews your project conventions and the action API, confirms the action's behavior and properties with you, then emits the action class, the optional properties model with form-component annotations, and the assembly-level `RegisterAutomationAction<>` registration. Full instructions are available in the [README](./plugins/automation-implementation/README.md).
+AI-assisted implementation of [Automation process](https://docs.kentico.com/documentation/business-users/digital-marketing/automation) extensions in Xperience by Kentico — starting with **custom automation actions** (custom step types in the Automation Builder). The AI reviews your project conventions and the action API, confirms the action's behavior and properties with you, then emits the action class, the optional properties model with form-component annotations, and the assembly-level `RegisterAutomationAction<>` registration. Full instructions are available in the [README](./plugins/kentico-digital-experience/README.md).
 
 | Skill                       | Description                                                                                                  |
 | --------------------------- | ------------------------------------------------------------------------------------------------------------ |
@@ -121,7 +121,7 @@ This repository is an [agent plugin marketplace](https://code.visualstudio.com/d
 
 ```bash
 copilot plugin marketplace add Kentico/xperience-by-kentico-kenticopilot
-copilot plugin install automation-implementation@xperience-by-kentico-kenticopilot
+copilot plugin install kentico-digital-experience@xperience-by-kentico-kenticopilot
 copilot plugin install widget-creation@xperience-by-kentico-kenticopilot
 copilot plugin install kx13-content-audit@xperience-by-kentico-kenticopilot
 copilot plugin install kx13-content-migration@xperience-by-kentico-kenticopilot
@@ -133,7 +133,7 @@ copilot plugin install configure-cd-repository@xperience-by-kentico-kenticopilot
 
 ```bash
 /plugin marketplace add Kentico/xperience-by-kentico-kenticopilot
-/plugin install automation-implementation@xperience-by-kentico-kenticopilot
+/plugin install kentico-digital-experience@xperience-by-kentico-kenticopilot
 /plugin install widget-creation@xperience-by-kentico-kenticopilot
 /plugin install kx13-content-audit@xperience-by-kentico-kenticopilot
 /plugin install kx13-content-migration@xperience-by-kentico-kenticopilot
