+++ plugins/kentico-digital-experience/.mcp.json
