+++ plugins/kentico-digital-experience/skills/automation-action-create/examples/SendContactSmsAction.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using CMS.Automation;
 using CMS.ContactManagement;
 
-using Kentico.Xperience.Admin.Base.FormAnnotations;
+using Kentico.Xperience.Admin.Base;
 
 using Microsoft.Extensions.Logging;
 
@@ -13,34 +13,12 @@ using Twilio.Rest.Api.V2010.Account;
 using Twilio.Types;
 
 [assembly: RegisterAutomationAction<SendContactSmsAction>(
-    identifier: "DancingGoat.SendContactSms",
+    identifier: SendContactSmsAction.IDENTIFIER,
     displayName: "Send SMS to contact",
-    IconName = "xp-message",
-    Tooltip = "Sends a templated SMS to the contact's mobile phone via Twilio.")]
+    IconName = Icons.Message,
+    Description = "Sends a templated SMS to the contact's mobile phone via Twilio.")]
 
 namespace Kentico.Xperience.DancingGoat.Automation;
-
-
-/// <summary>
-/// Configurable properties for <see cref="SendContactSmsAction"/>.
-/// </summary>
-public class SendContactSmsActionProperties : IAutomationActionProperties
-{
-    [TextInputComponent(
-        Label = "Sender ID",
-        ExplanationText = "Twilio \"from\" number in E.164 format (or an alphanumeric sender ID where supported).",
-        Order = 1)]
-    [RequiredValidationRule]
-    public string SenderId { get; set; } = "";
-
-
-    [TextAreaComponent(
-        Label = "Message template",
-        ExplanationText = "Use {ContactFirstName} as a placeholder for the contact's first name.",
-        Order = 2)]
-    [RequiredValidationRule]
-    public string MessageTemplate { get; set; } = "Hi {ContactFirstName}, thanks for being a Dancing Goat customer!";
-}
 
 
 /// <summary>
@@ -70,18 +48,15 @@ public class SendContactSmsAction(
     ILogger<SendContactSmsAction> logger)
     : AutomationAction<SendContactSmsActionProperties>
 {
+    public const string IDENTIFIER = "DancingGoat.SendContactSms";
+
+
     public override async Task Execute(
         SendContactSmsActionProperties properties,
         AutomationProcessContext context,
         CancellationToken cancellationToken)
     {
-        if (context.ProcessedObject is not ContactInfo contact)
-        {
-            logger.LogWarning(
-                "Skipping SendContactSms — processed object is not a contact (got '{ObjectType}').",
-                context.ProcessedObject?.TypeInfo.ObjectType);
-            return;
-        }
+        ContactInfo contact = await context.GetProcessedObject(cancellationToken);
 
         var phone = ResolveContactPhone(contact);
         if (string.IsNullOrWhiteSpace(phone))
