+++ plugins/kentico-digital-experience/skills/automation-action-create/examples/UpdateContactConsentAction.cs
@@ -7,38 +7,17 @@ using CMS.ContactManagement;
 using CMS.DataEngine;
 using CMS.DataProtection;
 
-using Kentico.Xperience.Admin.Base.FormAnnotations;
+using Kentico.Xperience.Admin.Base;
 
 using Microsoft.Extensions.Logging;
 
 [assembly: RegisterAutomationAction<UpdateContactConsentAction>(
-    identifier: "DancingGoat.UpdateContactConsent",
+    identifier: UpdateContactConsentAction.IDENTIFIER,
     displayName: "Update contact consent",
-    IconName = "xp-check-circle",
-    Tooltip = "Grants or revokes the selected consent for the contact via IConsentAgreementService.")]
+    IconName = Icons.CheckCircle,
+    Description = "Grants or revokes the selected consent for the contact via IConsentAgreementService.")]
 
 namespace Kentico.Xperience.DancingGoat.Automation;
-
-
-/// <summary>
-/// Configurable properties for <see cref="UpdateContactConsentAction"/>.
-/// </summary>
-public class UpdateContactConsentActionProperties : IAutomationActionProperties
-{
-    [ObjectIdSelectorComponent(
-        ConsentInfo.OBJECT_TYPE,
-        Label = "Consent",
-        Order = 1)]
-    [RequiredValidationRule]
-    public int? Consent { get; set; }
-
-
-    [DropDownComponent(
-        Label = "Action",
-        Options = "Agree;Grant consent\nRevoke;Revoke consent",
-        Order = 2)]
-    public string Action { get; set; } = UpdateContactConsentAction.AgreeAction;
-}
 
 
 /// <summary>
@@ -51,34 +30,30 @@ public class UpdateContactConsentAction(
     ILogger<UpdateContactConsentAction> logger)
     : AutomationAction<UpdateContactConsentActionProperties>
 {
+    public const string IDENTIFIER = "DancingGoat.UpdateContactConsent";
+
     internal const string AgreeAction = "Agree";
     internal const string RevokeAction = "Revoke";
 
 
-    public override Task Execute(
+    public override async Task Execute(
         UpdateContactConsentActionProperties properties,
         AutomationProcessContext context,
         CancellationToken cancellationToken)
     {
-        if (context.ProcessedObject is not ContactInfo contact)
-        {
-            logger.LogWarning(
-                "Skipping UpdateContactConsent — processed object is not a contact (got '{ObjectType}').",
-                context.ProcessedObject?.TypeInfo.ObjectType);
-            return Task.CompletedTask;
-        }
+        ContactInfo contact = await context.GetProcessedObject(cancellationToken);
 
         if (properties.Consent is not int consentId)
         {
             logger.LogWarning("Skipping UpdateContactConsent for contact {ContactId} — no consent selected on the step.", contact.ContactID);
-            return Task.CompletedTask;
+            return;
         }
 
         var consent = consentInfoProvider.Get(consentId);
         if (consent is null)
         {
             logger.LogWarning("Skipping UpdateContactConsent for contact {ContactId} — consent {ConsentId} not found.", contact.ContactID, consentId);
-            return Task.CompletedTask;
+            return;
         }
 
         if (string.Equals(properties.Action, RevokeAction, StringComparison.OrdinalIgnoreCase))
@@ -91,7 +66,5 @@ public class UpdateContactConsentAction(
             consentAgreementService.Agree(contact, consent);
             logger.LogInformation("Granted consent '{ConsentName}' for contact {ContactId}.", consent.ConsentName, contact.ContactID);
         }
-
-        return Task.CompletedTask;
     }
 }
